# delivery/docs/swagger.yaml
basePath: /
definitions:
  model.Order:
    properties:
      address:
        type: string
      cost:
        type: integer
      id:
        type: integer
      order_status:
        $ref: '#/definitions/model.OrderStatus'
      payment_status:
        type: boolean
      payment_type:
        $ref: '#/definitions/model.PaymentType'
    type: object
  model.OrderStatus:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  model.PaymentType:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Delivery microservice's API for CRM
  title: impl-crm-delivery
  version: "1.0"
paths:
...
  /order-statuses/status/{id}:
    get:
      consumes:
      - application/json
      description: Returns order status by ID
      operationId: find-order-status-by-id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OrderStatus'
        "400":
          description: ""
        "500":
          description: ""
      summary: FindOrderStatusById
      tags:
      - OrderStatus
...
  /orders/order/{id}:
    get:
      consumes:
      - application/json
      description: Returns order info by ID
      operationId: find-orders-by-id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Order'
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: FindOrderById
      tags:
      - Order
...
swagger: "2.0"